+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.4
+
+- Will sort output based on image dimensions
+- Added a getBest methods
+
 # 1.0.3
 
 - Don't rely on HTTP 200 OK for predefined URLs, but check file size too